Split video export internals
